### 2018-12-07 #1122 @whedon
[PRE REVIEW]: pytransform3d: 3D Transformations for Python
**Submitting author:** @AlexanderFabisch (<a href="http://orcid.org/0000-0003-2824-7956">Alexander Fabisch</a>)
**Repository:** <a href="https://github.com/rock-learning/pytransform3d" target ="_blank">https://github.com/rock-learning/pytransform3d</a>
**Version:** 1.0
