+++ paper/paper.bib
@@ -12,7 +12,7 @@ ISSN={2296-9144}
 
 @Misc{SciPyLectureNotes,
 title={Scipy Lecture Notes},
-author={Ga{\"e}l Varoquaux and Emmanuelle Gouillart and Olav Vahtras}
+author={Ga{\"e}l Varoquaux and Emmanuelle Gouillart and Olav Vahtras},
 url="https://www.scipy-lectures.org/",
 year=2018,
 note = {[Online; accessed <3 November 2018>]}
