### 2018-12-08 7e91f14
Fix paper layout
