### 2018-12-08 879e370
Correct figure captions
