### 2019-01-19 430be80
Add unit test and doc for pydot feature
