+++ doc/source/transform_manager.rst
@@ -11,6 +11,15 @@ end-effector in the object's frame so that we can control it.
 .. plot:: ../../examples/plot_transform_manager.py
     :include-source:
 
+We can also export the underlying graph structure as a PNG with::
+
+    tm.write_png(filename)
+
+.. image:: _static/graph.png
+    :width: 50%
+    :align: center
+
+
 A subclass of the transformation manager is the
 :class:`~pytransform3d.urdf.UrdfTransformManager` which can be used to load
 robot definitions from URDF files. An example with a simple robot can be seen
