+++ .travis.yml
@@ -11,7 +11,7 @@ branches:
 install:
   - sudo apt-get update -qq
   - sudo apt-get install -qq liblapack-dev libopenblas-dev
-  - pip install numpy scipy lxml bs4 matplotlib coverage
+  - pip install numpy scipy lxml bs4 matplotlib coverage pydot
 script: "nosetests -sv ."
 notifications:
   recipients:
