+++ requirements.txt
@@ -1,4 +1,6 @@
 numpy
 scipy
+matplotlib
 lxml
 bs4
+pydot
