+++ requirements.txt
@@ -3,4 +3,4 @@ scipy
 matplotlib
 lxml
 bs4
-pydot
+pydot[all]
