+++ .circleci/config.yml
@@ -27,6 +27,7 @@ jobs:
             . venv/bin/activate
             pip install -r requirements.txt
             pip install nose
+            apt install graphviz
 
       - save_cache:
           paths:
