+++ doc/source/transform_manager.rst
@@ -6,7 +6,9 @@ It is sometimes very difficult to have an overview of all the transformations
 that are required to obtain one specific transformation. Suppose you have
 a robot with a camera that can observe the robot's end-effector and an object
 that we want to manipulate. We would like to know the position of the
-end-effector in the object's frame so that we can control it.
+end-effector in the object's frame so that we can control it. The
+:class:`~pytransform3d.transform_manager.TransformManager` can handle this
+for you.
 
 .. plot:: ../../examples/plot_transform_manager.py
     :include-source:
