### 2019-01-24 #1159 @luizirber
Hello @AlexanderFabisch! Great package =]

Some comments:

## License

It almost matches https://opensource.org/licenses/BSD-3-Clause
but is missing the copyright line in the beginning and has some minor format
differences. I suggest copying the license from the link and saving to your LICENSE file (GitHub will also be able to pick it up and show in the landing repo page).

Also fix https://github.com/rock-learning/pytransform3d/blob/879e37069e16e372ad10762a5bb32bf5256fb0de/setup.py#L25
https://docs.python.org/3/distutils/setupscript.html#additional-meta-data

## Installation

You can specify your dependencies on setup.py. It seems like you're almost there,
but the right key is `install_requires`, and not `requires`:
https://github.com/rock-learning/pytransform3d/blob/879e37069e16e372ad10762a5bb32bf5256fb0de/setup.py#L26

More info: https://packaging.python.org/discussions/install-requires-vs-requirements/

## Community guidelines

You can also create a CONTRIBUTING file (GitHub recognizes it and render the content when a new issue or PR is opened):
https://github.blog/2012-09-17-contributing-guidelines/
