### 2019-01-25 e6b9031
Rename New BSD to BSD-3-Clause
