+++ requirements.txt
@@ -2,5 +2,5 @@ numpy
 scipy
 matplotlib
 lxml
-bs4
+beautifulsoup4
 pydot[all]
