### 2019-01-25 545243d
Add CONTRIBUTING.md
