### 2019-01-25 8f03ecd
Fixes of contribution guidelines
