### 2019-01-25 35765e3
Further fixes of contribution guidelines
