### 2019-01-25 9d00c03
Point to sklearn's guidelines
