### 2019-01-25 #1159 @AlexanderFabisch
Hi @luizirber,

thanks for the kind words and the review! I hope I addressed everything in [this merge request](https://github.com/rock-learning/pytransform3d/pull/25).
