### 2019-01-28 #1159 @luizirber
thanks @AlexanderFabisch! Last recommendation: tag a new release for the software and upload to PyPI =]

@brainstorm I checked all the boxes, good to go!
