### 2019-01-29 #1159 @AlexanderFabisch
> thanks @AlexanderFabisch! Last recommendation: tag a new release for the software and upload to PyPI =]

Yep, 1.1 is available. Thanks!
