### 2019-01-29 c36cbb4
Correct bibliography
