+++ paper/paper.bib
@@ -79,10 +79,8 @@ author={Vyas Ramasubramani and Sharon C. Glotzer},
 title={rowan: A Python package for working with quaternions},
 journal={Journal of Open Source Software},
 year=2018,
-dor={10.21105/joss.00787},
+doi={10.21105/joss.00787},
 url={https://doi.org/10.21105/joss.00787},
 volume=3,
 issue=27
 }
-
-Ramasubramani et al., (2018). rowan: A Python package for working with quaternions . Journal of Open Source Software, 3(27), 787, https://doi.org/10.21105/joss.00787
