### 2019-01-29 #1159 @AlexanderFabisch
OK, the bibtex entry had a `dor`. I changed it to `doi`. :laughing:
