### 2019-01-30 #1159 @brainstorm
LGTM, @AlexanderFabisch can you deposit the release to zenodo and come back with the DOI?
