### 2019-01-30 #1159 @AlexanderFabisch
Yes, this is the DOI: 10.5281/zenodo.2553451
