### 2019-01-30 #1159 @danielskatz
Hi @brainstorm - We've moved to rotating editors-in-chief, and this week it's me. For the future, please notify @joss-eics when there's a paper ready to go, or to otherwise get the person on-duty...
